+++ app/jobs/export/backup_mitglieder_export_job.rb
@@ -3,28 +3,26 @@
 #  Copyright (c) 2024, Schweizer Alpen-Club. This file is part of
 #  hitobito_sac_cas and licensed under the Affero General Public License version 3
 #  or later. See the COPYING file at the top-level directory or at
-#  https://github.com/hitobito/hitobito_sac_cas
+#  https://github.com/hitobito/hitobito_sac_cas.
 
-class Export::BackupMitgliederExportJob < RecurringJob
-
-  run_every 1.day
+class Export::BackupMitgliederExportJob < BaseJob
+  self.parameters = [:group_id]
   self.use_background_job_logging = true
 
-  ROLE_TYPES_TO_BACKUP = [::Group::Sektion, ::Group::Ortsgruppe].freeze
-
-  def initialize
-    super
+  def initialize(group_id)
+    super()
+    @group = Group.find(group_id)
     @errors = []
   end
 
-  def perform_internal
-    relevant_groups.find_each do |group|
-      BackupMitgliederExport.new(group, sftp).call
-    rescue StandardError => e
-      error(self, e, group: group)
-      @errors << [group.id, e]
-      next
-    end
+  def perform
+    sftp.create_remote_dir(root_folder_path) unless sftp.directory?(root_folder_path)
+    sftp.create_remote_dir(folder_path) unless sftp.directory?(folder_path)
+
+    sftp.upload_file(csv, file_path)
+  rescue StandardError => e
+    error(self, e, group: @group)
+    @errors << [@group.id, e]
   end
 
   def log_results
@@ -35,8 +33,11 @@ class Export::BackupMitgliederExportJob < RecurringJob
 
   private
 
-  def relevant_groups
-    Group.where(type: ROLE_TYPES_TO_BACKUP.map(&:sti_name))
+  def csv
+    @csv ||= begin
+               user_id = nil
+               SacCas::Export::MitgliederExportJob.new(user_id, @group.id).data
+             end
   end
 
   def sftp
@@ -47,7 +48,16 @@ class Export::BackupMitgliederExportJob < RecurringJob
     Settings.sftp.config
   end
 
-  def next_run
-    interval.from_now.midnight + 5.minutes
+  def file_path
+    "#{folder_path}Adressen_#{@group.navision_id_padded}.csv"
   end
+
+  def folder_path
+    "#{root_folder_path}#{@group.navision_id}/"
+  end
+
+  def root_folder_path
+    'sektionen/'
+  end
+
 end
