+++ lib/hitobito_sac_cas/wagon.rb
@@ -22,7 +22,7 @@ module HitobitoSacCas
 
     config.to_prepare do # rubocop:disable Metrics/BlockLength
       JobManager.wagon_jobs += [
-        Export::BackupMitgliederExportJob,
+        Export::BackupMitgliederScheduleJob,
         PromoteNeuanmeldungenJob,
         Event::CloseApplicationsJob,
         Roles::TerminateTourenleiterJob
